requirements.txt
ultralytics
openai
opencv-python
asgiref==3.8.1
attrs==24.2.0
bleach==6.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-oauth-toolkit==3.0.1
django-password-validators==1.7.3
django-resized==1.0.3
djangorestframework==3.15.2
drf-api-logger==1.1.16
drf-social-oauth2==3.1.0
google-auth==2.36.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jwcrypto==1.5.6
loguru==0.7.3
lxml==5.3.0
MarkupSafe==3.0.2
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.2.1
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.15.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.8.2
